Experiment with grouping update transactions together
The idea is to group non-crucial updates together and apply them on the
database all in one transactions in order to increase throughput of the
server.

Used for such things as job and task status, things which can't be reliably
known in the case of power outage and other types of failures.